working on the login and signup or auth component

+++ client/src/components/Auth.js
@@ -1,27 +1,46 @@
 import React, { useState } from 'react';
-import { Form } from 'react-bootstrap';		// utilize instead!
+import { Container, Tab, Tabs } from 'react-bootstrap';
+import Login from './Login';
+import Signup from './Signup';
 import axios from 'axios';
+import '../styles/container.css';
 
 const Auth = () => {
 	const [state, setState] = useState({});
+	const [errorMessage, setErrorMessage] = useState('');
 
 	const clearInputs = () => {
-		document.getElementById('email').value = '';
-		document.getElementById('password').value = '';
+		document.getElementById('login-email').value = '';
+		document.getElementById('login-password').value = '';
+		document.getElementById('signup-email').value = '';
+		document.getElementById('signup-password').value = '';
+		document.getElementById('confirm-email').value = '';
+		document.getElementById('confirm-password').value = '';
+		setErrorMessage('');
 		setState({});
 	}
 
 	const handleChange = (e) => {
 		e.preventDefault();
+		console.log(e.target.value);
 		setState({...state, [e.target.id]: e.target.value });
 	}
 
-	const register = async (e) => {
+	const signup = async (e) => {
 		e.preventDefault();
+
+		if(state['signup-email'].normalize() !== state['confirm-email'].normalize() || state['signup-password'].normalize() !== state['confirm-password'].normalize()){
+			setErrorMessage('Please check if your email or password is typed in correctly');
+			return;
+		}
+
 		await axios
-			.post('/api/users', state)
+			.post('/api/users', {
+				email: state['signup-email'],
+				password: state['signup-password']
+			})
 			.then(response => {
-				console.log('Register response: ', response);
+				console.log('Signup response: ', response);
 				localStorage.setItem('x-auth-token', response.headers['x-auth-token']);
 				clearInputs();
 			})
@@ -31,7 +50,10 @@ const Auth = () => {
 	const login = async (e) => {
 		e.preventDefault();
 		await axios
-			.post('/api/login', state)
+			.post('/api/login', {
+				email: state['login-email'],
+				password: state['login-password']
+			})
 			.then(response => {
 				console.log('Login Response: ', response);
 				localStorage.setItem('x-auth-token', response.headers['x-auth-token']);
@@ -43,20 +65,17 @@ const Auth = () => {
 	return(
 		<>
 			<h1>Auth Page</h1>
-			<form>
-				<label>
-					<p>Email</p>
-					<input type="email" id="email" onChange={handleChange} />
-				</label>
-				<label>
-					<p>Password</p>
-					<input type="password" id="password" onChange={handleChange} />
-				</label>
-				<div>
-					<button type="submit" onClick={register}>Register</button>
-					<button type="submit" onClick={login}>Login</button>
-				</div>
-			</form>
+			<Container className="border rounded border-secondary auth-container">
+				<Tabs defaultActiveKey="login" id="uncontrolled-tab-example" className="mb-3" onSelect={clearInputs}>
+					<Tab eventKey="login" title="Login">
+						<Login handleChange={handleChange} login={login} />
+					</Tab>
+					<Tab eventKey="signup" title="Signup">
+					<Signup handleChange={handleChange} signup={signup} />
+					</Tab>
+				</Tabs>
+				<h6>{errorMessage}</h6>
+			</Container>
 		</>
 	);
 }
